refactor(project): migrate to typescript
