+++ .github/workflows/main.yml
@@ -11,7 +11,6 @@ jobs:
       matrix:
         node-version:
           - 14
-          - 12
     steps:
       - uses: actions/checkout@v2
       - uses: actions/setup-node@v1
