+++ .gitignore
@@ -1,4 +1,4 @@
-node_modules
+distribution/
+node_modules/
+package-lock.json
 yarn.lock
-.nyc_output
-coverage
