+++ package.json
@@ -11,16 +11,20 @@
 		"url": "https://sindresorhus.com"
 	},
 	"type": "module",
-	"exports": "./index.js",
+	"exports": "./distribution/index.js",
+	"types": "./distribution/index.d.ts",
 	"engines": {
 		"node": ">=12"
 	},
 	"scripts": {
-		"test": "xo && ava && tsd"
+		"test": "xo && npm run build && ava",
+		"debug": "PWDEBUG=1 ava --timeout=2m",
+		"release": "np",
+		"build": "del-cli distribution/ && tsc --project tsconfig.dist.json",
+		"prepare": "npm run build"
 	},
 	"files": [
-		"index.js",
-		"index.d.ts"
+		"distribution"
 	],
 	"keywords": [
 		"fetch",
@@ -45,18 +49,27 @@
 		"node-fetch"
 	],
 	"devDependencies": {
+		"@sindresorhus/tsconfig": "^1.0.1",
+		"@types/body-parser": "^1.19.0",
+		"@types/busboy": "^0.2.3",
+		"@types/express": "^4.17.11",
+		"@types/node-fetch": "^2.5.8",
+		"@types/pify": "^5.0.0",
 		"abort-controller": "^3.0.0",
 		"ava": "^3.15.0",
-		"body": "^5.1.0",
+		"body-parser": "^1.19.0",
 		"busboy": "^0.3.1",
-		"create-test-server": "2.1.1",
-		"delay": "^4.1.0",
-		"form-data": "^3.0.0",
-		"node-fetch": "^2.5.0",
-		"nyc": "^15.0.0",
-		"playwright-chromium": "^1.7.1",
-		"tsd": "^0.13.1",
-		"xo": "^0.25.3"
+		"del-cli": "^3.0.1",
+		"delay": "^5.0.0",
+		"express": "^4.17.1",
+		"form-data": "^4.0.0",
+		"node-fetch": "^2.6.1",
+		"pify": "^5.0.0",
+		"playwright-chromium": "^1.10.0",
+		"raw-body": "^2.4.1",
+		"ts-node": "^9.1.1",
+		"typescript": "^4.2.3",
+		"xo": "^0.38.2"
 	},
 	"sideEffects": false,
 	"xo": {
@@ -67,31 +80,40 @@
 			"globalThis"
 		],
 		"rules": {
-			"import/no-useless-path-segments": "off",
-			"unicorn/import-index": "off",
-			"import/extensions": "off"
+			"@typescript-eslint/dot-notation": "off",
+			"ava/no-import-test-files": "off",
+			"unicorn/filename-case": "off"
 		}
 	},
 	"ava": {
+		"files": [
+			"test/*"
+		],
 		"require": [
-			"./test/_require"
+			"./test/_require.ts"
 		],
 		"nonSemVerExperiments": {
+			"configurableModuleFormat": true,
 			"nextGenConfig": true
-		}
-	},
-	"tsd": {
-		"compilerOptions": {
-			"lib": [
-				"es2019",
-				"dom"
-			]
-		}
+		},
+		"extensions": {
+			"ts": "module"
+		},
+		"nodeArguments": [
+			"--loader=ts-node/esm"
+		]
 	},
 	"nyc": {
 		"reporter": [
 			"text",
+			"html",
 			"lcov"
+		],
+		"extension": [
+			".ts"
+		],
+		"exclude": [
+			"**/test/**"
 		]
 	}
 }
