+++ source/utils/is.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line @typescript-eslint/ban-types
+export const isObject = (value: unknown): value is object => value !== null && typeof value === 'object';
