+++ source/utils/merge.ts
@@ -1,0 +1,59 @@
+import type {KyHeadersInit, Options} from '../types/options.js';
+import {isObject} from './is.js';
+
+export const validateAndMerge = (...sources: Array<Partial<Options> | undefined>): Partial<Options> => {
+	for (const source of sources) {
+		if ((!isObject(source) || Array.isArray(source)) && typeof source !== 'undefined') {
+			throw new TypeError('The `options` argument must be an object');
+		}
+	}
+
+	return deepMerge({}, ...sources);
+};
+
+export const mergeHeaders = (source1: KyHeadersInit = {}, source2: KyHeadersInit = {}) => {
+	const result = new globalThis.Headers(source1 as HeadersInit);
+	const isHeadersInstance = source2 instanceof globalThis.Headers;
+	const source = new globalThis.Headers(source2 as HeadersInit);
+
+	for (const [key, value] of source.entries()) {
+		if ((isHeadersInstance && value === 'undefined') || value === undefined) {
+			result.delete(key);
+		} else {
+			result.set(key, value);
+		}
+	}
+
+	return result;
+};
+
+export const deepMerge = <T>(...sources: Array<Partial<T> | undefined>): T => {
+	let returnValue: any = {};
+	let headers = {};
+
+	for (const source of sources) {
+		if (Array.isArray(source)) {
+			if (!Array.isArray(returnValue)) {
+				returnValue = [];
+			}
+
+			returnValue = [...returnValue, ...source];
+		} else if (isObject(source)) {
+			for (let [key, value] of Object.entries(source)) {
+				if (isObject(value) && key in returnValue) {
+					value = deepMerge(returnValue[key], value);
+				}
+
+				returnValue = {...returnValue, [key]: value};
+			}
+
+			if (isObject((source as any).headers)) {
+				headers = mergeHeaders(headers, (source as any).headers);
+			}
+		}
+
+		returnValue.headers = headers;
+	}
+
+	return returnValue;
+};
