+++ test/_require.ts
@@ -2,9 +2,14 @@ import fetch, {Headers, Request, Response} from 'node-fetch';
 import AbortController from 'abort-controller';
 import FormData from 'form-data';
 
-globalThis.fetch = fetch;
-globalThis.Headers = Headers;
-globalThis.Request = Request;
-globalThis.Response = Response;
 globalThis.AbortController = AbortController;
+// @ts-expect-error
+globalThis.fetch = fetch;
+// @ts-expect-error
+globalThis.Headers = Headers;
+// @ts-expect-error
+globalThis.Request = Request;
+// @ts-expect-error
+globalThis.Response = Response;
+// @ts-expect-error
 globalThis.FormData = FormData;
