+++ test/fetch.ts
@@ -1,10 +1,12 @@
 import test from 'ava';
-import ky from '../index.js';
+import ky from '../source/index.js';
 
 test.serial('relative URLs are passed to fetch unresolved', async t => {
 	const originalFetch = globalThis.fetch;
 	globalThis.fetch = async input => {
+		// @ts-expect-error @TODO
 		t.true(input.url.startsWith('/'));
+		// @ts-expect-error @TODO
 		return new Response(input.url);
 	};
 
@@ -20,15 +22,40 @@ test.serial('relative URLs are passed to fetch unresolved', async t => {
 test('fetch option takes a custom fetch function', async t => {
 	t.plan(12);
 
-	const customFetch = async input => {
+	const customFetch: typeof fetch = async input => {
 		t.true(input instanceof Request);
+		// @ts-expect-error @TODO
 		return new Response(input.url);
 	};
 
 	t.is(await ky('/unicorn', {fetch: customFetch}).text(), '/unicorn');
-	t.is(await ky('/unicorn', {fetch: customFetch, searchParams: {foo: 'bar'}}).text(), '/unicorn?foo=bar');
-	t.is(await ky('/unicorn#hash', {fetch: customFetch, searchParams: 'foo'}).text(), '/unicorn?foo#hash');
-	t.is(await ky('/unicorn?old', {fetch: customFetch, searchParams: 'new'}).text(), '/unicorn?new');
-	t.is(await ky('/unicorn?old#hash', {fetch: customFetch, searchParams: 'new'}).text(), '/unicorn?new#hash');
+	t.is(
+		await ky('/unicorn', {
+			fetch: customFetch,
+			searchParams: {foo: 'bar'}
+		}).text(),
+		'/unicorn?foo=bar'
+	);
+	t.is(
+		await ky('/unicorn#hash', {
+			fetch: customFetch,
+			searchParams: 'foo'
+		}).text(),
+		'/unicorn?foo#hash'
+	);
+	t.is(
+		await ky('/unicorn?old', {
+			fetch: customFetch,
+			searchParams: 'new'
+		}).text(),
+		'/unicorn?new'
+	);
+	t.is(
+		await ky('/unicorn?old#hash', {
+			fetch: customFetch,
+			searchParams: 'new'
+		}).text(),
+		'/unicorn?new#hash'
+	);
 	t.is(await ky('unicorn', {fetch: customFetch, prefixUrl: '/api/'}).text(), '/api/unicorn');
 });
