--- test/helpers/disable-abort-controller.js
@@ -1,1 +1,0 @@
-window.AbortController = undefined;
