--- test/helpers/disable-stream-support.js
@@ -1,1 +1,0 @@
-window.ReadableStream = undefined;
