+++ test/http-error.ts
@@ -1,14 +1,15 @@
 import test from 'ava';
-import {Response} from 'node-fetch';
-import ky from '../index.js';
+import ky from '../source/index.js';
 
-function createFakeResponse({status, statusText}) {
+function createFakeResponse({status, statusText}: {status?: number; statusText?: string}): Response {
 	// Start with a realistic fetch Response.
 	const response = {...new Response()};
 
 	// Assign these values because the Response() constructor doesn't
 	// support setting them to undefined.
+	// @ts-expect-error
 	response.status = status;
+	// @ts-expect-error
 	response.statusText = statusText;
 
 	return response;
@@ -16,6 +17,7 @@ function createFakeResponse({status, statusText}) {
 
 test('HTTPError handles undefined response.statusText', t => {
 	const status = 500;
+	// @ts-expect-error missing Request
 	const error = new ky.HTTPError(
 		// This simulates the case where a browser Response object does
 		// not define statusText, such as IE, Safari, etc.
@@ -27,6 +29,7 @@ test('HTTPError handles undefined response.statusText', t => {
 });
 
 test('HTTPError handles undefined response.status', t => {
+	// @ts-expect-error missing Request
 	const error = new ky.HTTPError(
 		// This simulates a catastrophic case where some unexpected
 		// response object was sent to HTTPError.
@@ -37,6 +40,7 @@ test('HTTPError handles undefined response.status', t => {
 });
 
 test('HTTPError handles a response.status of 0', t => {
+	// @ts-expect-error missing Request
 	const error = new ky.HTTPError(
 		// Apparently, it's possible to get a response status of 0.
 		createFakeResponse({statusText: undefined, status: 0})
