+++ test/prefix-url.ts
@@ -1,17 +1,21 @@
 import test from 'ava';
-import createTestServer from 'create-test-server';
-import ky from '../index.js';
+import ky from '../source/index.js';
+import {createHttpTestServer} from './helpers/create-http-test-server.js';
 
 test('prefixUrl option', async t => {
-	const server = await createTestServer();
-	server.get('/', (request, response) => {
+	const server = await createHttpTestServer();
+	server.get('/', (_request, response) => {
 		response.end('zebra');
 	});
-	server.get('/api/unicorn', (request, response) => {
+	server.get('/api/unicorn', (_request, response) => {
 		response.end('rainbow');
 	});
 
-	t.is(await ky(`${server.url}/api/unicorn`, {prefixUrl: false}).text(), 'rainbow');
+	t.is(
+		// @ts-expect-error {prefixUrl: boolean} isn't officially supported
+		await ky(`${server.url}/api/unicorn`, {prefixUrl: false}).text(),
+		'rainbow'
+	);
 	t.is(await ky(`${server.url}/api/unicorn`, {prefixUrl: ''}).text(), 'rainbow');
 	t.is(await ky(new URL(`${server.url}/api/unicorn`), {prefixUrl: ''}).text(), 'rainbow');
 	t.is(await ky('api/unicorn', {prefixUrl: server.url}).text(), 'rainbow');
@@ -23,11 +27,14 @@ test('prefixUrl option', async t => {
 	t.is(await ky('', {prefixUrl: `${server.url}/`}).text(), 'zebra');
 	t.is(await ky('', {prefixUrl: new URL(server.url)}).text(), 'zebra');
 
-	t.throws(() => {
-		ky('/unicorn', {prefixUrl: `${server.url}/api`});
-	}, {
-		message: '`input` must not begin with a slash when using `prefixUrl`'
-	});
+	t.throws(
+		() => {
+			void ky('/unicorn', {prefixUrl: `${server.url}/api`});
+		},
+		{
+			message: '`input` must not begin with a slash when using `prefixUrl`'
+		}
+	);
 
 	await server.close();
 });
