+++ test/retry.ts
@@ -1,6 +1,6 @@
 import test from 'ava';
-import createTestServer from 'create-test-server';
-import ky from '../index.js';
+import ky from '../source/index.js';
+import {createHttpTestServer} from './helpers/create-http-test-server.js';
 
 const fixture = 'fixture';
 const defaultRetryCount = 2;
@@ -10,8 +10,8 @@ const lastTried413access = Date.now();
 test('network error', async t => {
 	let requestCount = 0;
 
-	const server = await createTestServer();
-	server.get('/', (request, response) => {
+	const server = await createHttpTestServer();
+	server.get('/', (_request, response) => {
 		requestCount++;
 
 		if (requestCount === defaultRetryCount) {
@@ -29,8 +29,8 @@ test('network error', async t => {
 test('status code 500', async t => {
 	let requestCount = 0;
 
-	const server = await createTestServer();
-	server.get('/', (request, response) => {
+	const server = await createHttpTestServer();
+	server.get('/', (_request, response) => {
 		requestCount++;
 
 		if (requestCount === defaultRetryCount) {
@@ -48,8 +48,8 @@ test('status code 500', async t => {
 test('only on defined status codes', async t => {
 	let requestCount = 0;
 
-	const server = await createTestServer();
-	server.get('/', (request, response) => {
+	const server = await createHttpTestServer();
+	server.get('/', (_request, response) => {
 		requestCount++;
 
 		if (requestCount === defaultRetryCount) {
@@ -67,8 +67,8 @@ test('only on defined status codes', async t => {
 test('not on POST', async t => {
 	let requestCount = 0;
 
-	const server = await createTestServer();
-	server.post('/', (request, response) => {
+	const server = await createHttpTestServer();
+	server.post('/', (_request, response) => {
 		requestCount++;
 
 		if (requestCount === defaultRetryCount) {
@@ -78,7 +78,9 @@ test('not on POST', async t => {
 		}
 	});
 
-	await t.throwsAsync(ky.post(server.url).text(), {message: /Internal Server Error/});
+	await t.throwsAsync(ky.post(server.url).text(), {
+		message: /Internal Server Error/
+	});
 
 	await server.close();
 });
@@ -86,8 +88,8 @@ test('not on POST', async t => {
 test('respect 413 Retry-After', async t => {
 	let requestCount = 0;
 
-	const server = await createTestServer();
-	server.get('/', (request, response) => {
+	const server = await createHttpTestServer();
+	server.get('/', (_request, response) => {
 		requestCount++;
 
 		if (requestCount === defaultRetryCount) {
@@ -109,13 +111,14 @@ test('respect 413 Retry-After', async t => {
 test('respect 413 Retry-After with timestamp', async t => {
 	let requestCount = 0;
 
-	const server = await createTestServer();
-	server.get('/', (request, response) => {
+	const server = await createHttpTestServer({bodyParser: false});
+	server.get('/', (_request, response) => {
 		requestCount++;
 		if (requestCount === defaultRetryCount) {
 			response.end((Date.now() - lastTried413access).toString());
 		} else {
-			const date = (new Date(Date.now() + (retryAfterOn413 * 1000))).toUTCString();
+			// @NOTE we need to round up to the next second due to http-date resolution
+			const date = new Date(Date.now() + ((retryAfterOn413 + 1) * 1000)).toUTCString();
 			response.writeHead(413, {
 				'Retry-After': date
 			});
@@ -133,8 +136,8 @@ test('respect 413 Retry-After with timestamp', async t => {
 test('doesn\'t retry on 413 without Retry-After header', async t => {
 	let requestCount = 0;
 
-	const server = await createTestServer();
-	server.get('/', (request, response) => {
+	const server = await createHttpTestServer();
+	server.get('/', (_request, response) => {
 		requestCount++;
 		response.sendStatus(413);
 	});
@@ -150,8 +153,8 @@ test('doesn\'t retry on 413 without Retry-After header', async t => {
 test('respect number of retries', async t => {
 	let requestCount = 0;
 
-	const server = await createTestServer();
-	server.get('/', (request, response) => {
+	const server = await createHttpTestServer();
+	server.get('/', (_request, response) => {
 		requestCount++;
 		response.sendStatus(408);
 	});
@@ -174,13 +177,13 @@ test('respect number of retries', async t => {
 test('respect retry methods', async t => {
 	let requestCount = 0;
 
-	const server = await createTestServer();
-	server.post('/', (request, response) => {
+	const server = await createHttpTestServer();
+	server.post('/', (_request, response) => {
 		requestCount++;
 		response.sendStatus(408);
 	});
 
-	server.get('/', (request, response) => {
+	server.get('/', (_request, response) => {
 		requestCount++;
 		response.sendStatus(408);
 	});
@@ -219,8 +222,8 @@ test('respect retry methods', async t => {
 test('respect maxRetryAfter', async t => {
 	let requestCount = 0;
 
-	const server = await createTestServer();
-	server.get('/', async (request, response) => {
+	const server = await createHttpTestServer();
+	server.get('/', async (_request, response) => {
 		requestCount++;
 
 		response.writeHead(413, {
@@ -263,18 +266,15 @@ test('respect maxRetryAfter', async t => {
 test('retry - can provide retry as number', async t => {
 	let requestCount = 0;
 
-	const server = await createTestServer();
-	server.get('/', async (request, response) => {
+	const server = await createHttpTestServer();
+	server.get('/', async (_request, response) => {
 		requestCount++;
 		response.sendStatus(408);
 	});
 
-	await t.throwsAsync(
-		ky(server.url, {retry: 5}).text(),
-		{
-			message: /Request Timeout/
-		}
-	);
+	await t.throwsAsync(ky(server.url, {retry: 5}).text(), {
+		message: /Request Timeout/
+	});
 	t.is(requestCount, 5);
 
 	await server.close();
@@ -283,9 +283,9 @@ test('retry - can provide retry as number', async t => {
 test('doesn\'t retry on 413 with empty statusCodes and methods', async t => {
 	let requestCount = 0;
 
-	const server = await createTestServer();
+	const server = await createHttpTestServer();
 
-	server.get('/', async (request, response) => {
+	server.get('/', async (_request, response) => {
 		requestCount++;
 		response.sendStatus(413);
 	});
@@ -311,8 +311,8 @@ test('doesn\'t retry on 413 with empty statusCodes and methods', async t => {
 test('doesn\'t retry on 413 with empty methods', async t => {
 	let requestCount = 0;
 
-	const server = await createTestServer();
-	server.get('/', async (request, response) => {
+	const server = await createHttpTestServer();
+	server.get('/', async (_request, response) => {
 		requestCount++;
 		response.sendStatus(413);
 	});
@@ -337,8 +337,8 @@ test('doesn\'t retry on 413 with empty methods', async t => {
 test('does retry on 408 with methods provided as array', async t => {
 	let requestCount = 0;
 
-	const server = await createTestServer();
-	server.get('/', async (request, response) => {
+	const server = await createHttpTestServer();
+	server.get('/', async (_request, response) => {
 		requestCount++;
 		response.sendStatus(408);
 	});
@@ -363,8 +363,8 @@ test('does retry on 408 with methods provided as array', async t => {
 test('does retry on 408 with statusCodes provided as array', async t => {
 	let requestCount = 0;
 
-	const server = await createTestServer();
-	server.get('/', async (request, response) => {
+	const server = await createHttpTestServer();
+	server.get('/', async (_request, response) => {
 		requestCount++;
 		response.sendStatus(408);
 	});
@@ -389,8 +389,8 @@ test('does retry on 408 with statusCodes provided as array', async t => {
 test('doesn\'t retry when retry.limit is set to 0', async t => {
 	let requestCount = 0;
 
-	const server = await createTestServer();
-	server.get('/', (request, response) => {
+	const server = await createHttpTestServer();
+	server.get('/', (_request, response) => {
 		requestCount++;
 		response.sendStatus(408);
 	});
@@ -412,11 +412,12 @@ test('doesn\'t retry when retry.limit is set to 0', async t => {
 });
 
 test('throws when retry.methods is not an array', async t => {
-	const server = await createTestServer();
+	const server = await createHttpTestServer();
 
 	t.throws(() => {
-		ky(server.url, {
+		void ky(server.url, {
 			retry: {
+				// @ts-expect-error
 				methods: 'get'
 			}
 		});
@@ -426,11 +427,12 @@ test('throws when retry.methods is not an array', async t => {
 });
 
 test('throws when retry.statusCodes is not an array', async t => {
-	const server = await createTestServer();
+	const server = await createHttpTestServer();
 
 	t.throws(() => {
-		ky(server.url, {
+		void ky(server.url, {
 			retry: {
+				// @ts-expect-error
 				statusCodes: 403
 			}
 		});
